Create futures greedily
This means that when if we finish early, we'll have sent 16 queries off
to Cassandra unnecessarily (previously the expected wastage was 8).
However, we'll be more efficient overall, as we will likely have to
wait less for every 16th query to return.

This change also does away with a nested loop.