Split ExecutorInheritableThreadLocal out of commons-executors (#2961)
* Split ExecutorInheritableThreadLocal out of commons-executors

The introduction of remoting3:tracing transitively added guava, jackson,
etc. to the dependency footprint of commons-executors. This is painful
for api projects that don't want these dependencies.

ExecutorInheritableThreadLocal is useful for api projects to declare
shared thread locals, so it's useful for it to be split out into a
separate project with no dependencies.

* update version locks

* add release note, build.gradle comment