Be nice to Cassandra client's input buffers
Calling clear() on TMemoryInputTransport sets its 'buf' field to null
which can later result in a NullPointerException in places like
TBinaryProtocol.readI32().

Instead, we can call reset() with an empty byte array to achieve
a similar effect without using null pointers. For safety, we also do it
only if there are no bytes left in the buffer to avoid effectively
modifying its state. Not sure why there would be extra bytes left in the
buffer after a successful request, but the evidence suggests that it
might happen occasionally.