Puncher no longer uses range scans
One of the things we'd like to do is to have some usages of AtlasDB use
the Murmur partitioner in Cassandra. A consequence of this is that
traditional AtlasDB range scans will be a lot harder, and that we should
be using dynamic columns for this.

This PR moves to a scheme for the puncher where the same row is used
for some period of time (14 days as currently configured), and we range
scan within this row.

If the request is not within the current row, we look back one, and then
30.

Additionally, I removed the 'get wall clock time for timestamp' method,
leaving it in the KVS implementation only. Right now it's only used in
a cli, so I consider it dead code.