Harden the background sweeper against InterruptedException bugs
The current implementation relies on all code everywhere correctly handling thread interruption; if any wait call anywhere erroneously swallows the interrupt, the daemon will continue to run, preventing clean shutdown. Instead, trip a CountDownLatch after interrupting the thread, and replace all calls to `Thread.sleep` with a wait on this latch, so even if the interrupt gets swallowed, the sweeper will still terminate when it next tries to wait.

[no release notes]