Exclude swept time from SweepResults.equals
A few tests break because we create a couple of empty results, which
have System.currentTimeMillis as the started time, and then expect them
to equals() each other. Adding Value.Auxiliary excludes this field from
equals().

As a side-effect, the field is also excluded from toString(), but since
this class isn't serialised, we don't mind.

[no release notes]