Remove dropwizard-jackson from atlasdb-config dependencies
Many consumers of AtlasDB are not Dropwizard-based; by including
dropwizard-jackson and its transitive deps, such as dropwizard-util,
unnecessary classpath dependencies are introduced. Additionally,
it introduces hte opportunity for transitive version conflicts
if the consumer is dropwizard-based.

This change removes dropwizard-jackson as a dependency of atlasdb-config.
One consequence of this is that the Dropwizard DiscoverableSubtypeResolver
is no longer used when deserializing configuration; however, no references
to the Dropwizard discoverable annotation were found in the atlasdb repo.

If consumers relied on this functionality this may be a break; if so,
workarounds such as reimplementing DiscoverableSubtypeResolver are available.