Remove dropwizard-jackson from atlasdb-config dependencies
Many consumers of AtlasDB are not Dropwizard-based; by including
dropwizard-jackson and its transitive deps, such as dropwizard-util,
unnecessary classpath dependencies are introduced. Additionally,
it introduces hte opportunity for transitive version conflicts
if the consumer is dropwizard-based.

This change removes dropwizard-jackson as a dependency of atlasdb-config.
One consequence of this is that the Dropwizard DiscoverableSubtypeResolver
has been reimplemented and registered with the ObjectMapper used for
serialization.