[no release notes] Delete the CQL KVS and some other unused stuff
This is likely to be contentious, but the gist is:

1. The CQL stuff has never been used in production, so likely doesn't
work well enough.
2. There's no migration path; the CQL stuff likely can _never_ be used
in production. In particular, we're unlikely to move over in this way
(big-bang switch) - the only way to migrate is the way we're currently
doing it (migrating one call at a time, tactically).
3. There's almost no testing, as you can see there are no test changes
here. In AtlasDB, what this generally means is that there are loads of
bugs.
4. This code actually ends up getting modified a lot (100 times in the
last 3 years) so is clearly some kind of burden on the AtlasDB team.

Given that, would like to delete, and we can always resurrect later?