Use a daemon executor
Internally we've had a few cases recently where we've failed to shut
down properly. All the executors _we_ create are properly shut down
when a termination signal is seen, and this is the case almost all
of the time in AtlasDB, except in this place, which is brand new.

I've diffed our dependencies, and this seems like the most likely
place for our issues to show up.