Deletes for a single row are now Cassandra tombstones per row not per cell
I had been informed that we'd be deleting the whole row in one go, but
as I saw the number of cells read increase, I realised I'd seen no
evidence for this actually happening :P upon investigating, I realised
that it does not.

This PR makes the 'deleteRows' kvs method drop a tombstone for the whole
row if we're trying to delete a whole row. This avoids

It also trivially extends to DBKVS, but I haven't implemented it there
because the whole method isn't implemented there.