[no release notes] Less locking (#3356)
Locks are generally quite hard to reason about and also quite slow,
hence all the javadoc in this class and code to deal with the
concurrent ways of accessing the set.

This PR changes to use an ArrayBlockingQueue, which I believe to be
a better primitive here; e.g. it makes clear that all elements will
be seen eventually.