A dictionary for table names for use in targeted sweep
Even now, the overheads of targeted sweep can be relatively high on
systems with many indices and high ingestion rate (notably the lack of
block compression and the comparatively small disk sizes bite us).

This will require care wrt our internal backup infrastructure. I'm going
to coordinate this, but it adds an extra 'must repair these tables in C*
post restore' because this is one of those tables where even a small
inconsistency can become catastrophic.

Benchmarking the storage from this we can see that the improvement is
significant. At the moment in our benchmarking, the average
row-column-value total overhead is 80 bytes over the internal workload.

However, we can account for at most 19 bytes in the row and column, so
there must be on average 61 bytes in the value. Moreover, some samples
imply that a good chunk of the more written-to-tables have #bytes used
values >= 30 characters, which would imply that we could plausibly
see another step change improvement here.