[no release notes] Never use the 'getActualValues' method
DbKvs behaves differently to the other KVSes, and it's non-trivial to
unify the behaviour. I'm going to file the issue, but this is the
easiest workaround.

Explicitly, if you use CAS to do a PUE, all C* and InMemory will set the
'current values', whereas DbKVS will not. Changing this would require
some fairly exciting SQL.