Clean up TimeLockClient
This is a bug - the actual timestamp-and-lock close method was not getting registered as a close callback of its transaction manager. That close method closes `TimeLockClient`. Failing to do this causes thread leaks, especially in test scenarios where many transaction managers can be created and destroyed. I'm surprised the type checker allowed this bug to exist.