Use soft evict limits not evict limits with the pool
At the moment, we evict every connection that has been idle for 10
minutes every 10 minutes. If there is a min pool size (our default is
30) this is then immediately replaced with another SSL connection;
essentially just churning.

In practice, we've seen this cause issues internally around GC; SSL
connections filling up the old gen and then finalizer issues leading to
an ultimate GC overhead exceeded OOM.

I guess my concern would be that the Thrift connections actually break
after a certain period and so we are essentially evicting anyway (but
later, taking a performance hit); in this case we will see logging
occurring in such services indicating this.