CommitTsCache for targeted sweep expires entries after 5 minutes (#4197)
* Add generated changelog entries

* CommitTsCache for targeted sweep expires entries after 5 minutes

Right now, if you have e.g. 50 AtlasDB txn managers, the sweep commit
ts cache will hold onto 1MM start timestamps each, at a cost of about
40GB ram.

This code is also not instantiated in a dependency-injection friendly
fashion, and so it's not particularly fun to configure this.

I modify to use Caffeine (because Caffeine is better in every case than
Guava) and have a 5 minute expiry, to ensure that unused entries get
evicted and hopefully using much less ram in the long term.