Factor out timestamp cache
Internally, we have many AtlasDBs, each with their own timestamp cache.
We'd like to have one big one, with a single expiry policy - basically
because we think that it'll lead us to a more stable situation with less
GC-ing and OOMing.

Factor out an interface, and make it be pluggable.