A quick hack to see the viability of table level conflict handling (#4228)
* A quick hack to see the viability of table level conflict handling

This should represent ~= 0 overhead for our current users, and won't use
a huge amount of memory for anyone. That said, the code is kind of
crappy so I will probably refactor it.

Anyway, it's a try-out to get metrics as to the viability of:

What if timelock were able to do some very simple conflict checking
(e.g. check conflicts on tables or other very high level concepts only
(e.g. topics)). Would we be able to significantly reduce the number of
serializable transactions that get rolled back?

Obviously it's incomplete - if enabling serializable txns were lower
overhead, people would do it more. But it covers a good chunk of ground.

* Suppress checkstyle warnings

* More baseline