Rewrite disruptorautobatcher to just use a plain blocking queue
We've seen bugs internally caused by the autobatcher. In particular, the
code here:
https://github.com/LMAX-Exchange/disruptor/blob/master/src/main/java/com/lmax/disruptor/MultiProducerSequencer.java#L136
is causing a high context switch rate. Since I mostly used this class
because it seemed cute, and not because it seemed necessary, it seems
reasonable to rewrite it. This means that a few of the other classes
also become simpler.