Increase default buffer size to 2048 (#4535)
* Increase default buffer size to 8192

We see issues internally when the buffer fills up - we think there's a bug in the Disruptor where it spinloops on LockSupport.park(1 nanosecond) and death spirals. Thankfully, the amount of stuff in the queue is proportional to the number of threads that block on it and so this is typically at most in the low thousands. I therefore think that while my other PR will work fine, this is much less churn, will not significantly bloat memory, and will take us out of the woods.

* Add generated changelog entries