Start migrating AtlasDB to full conjure (#4545)
* Start migrating AtlasDB to full conjure

At present, we see some performance issues with timelock internally.
These will take some debugging, but one component is that our
recommendation for high-performance servers internally is not to use
Jersey (especially not jersey async, with the extra gubbins we have) but
is to use conjure-undertow.

We also view not using conjure defined APIs as AtlasDB tech debt.

Migrating Atlas in one go to conjure undertow is difficult if not
impossible. But there are two endpoints which represent most calls -
getLeaderTime and startTransactions.

This PR:

1. Adds a conjure API project (timelock-api) and defines a subset of the
AtlasDB timelock API in it.
2. Exposes this via the timelock service, with an Undertow and Jersey
implementation.
3. Uses the client for these new endpoints on the client side, provided
a boolean flag is set.

Right now, tests are a little fiddly because I'm coding on my home
laptop and they don't work at all there.

Places which are still potentially a little off:
1. Error handling logic in the Conjure undertow client is just
_different_ to Jersey's. But I think I caught all the special error
handlers.
2. We can't actually use this codepath until we can expose an internal
timelock dependency. So I've got a boolean flag that enables it
conditionally so we can test the timelock server implementation.
3. I think that the client might fall back to feign in some cases? And
this won't work if it calls the non-leader. Not sure how we'd want to
change that one.

* Delete redundant classes

* Imports

* Imorts

* Fix bug in jersey code

* Fix compile

* Mock correctly

* Add generated changelog entries

* Disable license checks for conjure generated code

* Make class public

* Don't let the TimelockResource own almost all strings.

As a consequence, 'tl' is no longer a valid name for an Atlas client.
But it's not a name at present.

* Move variable to end of line

* Remove repositories configuration from shared.gradle

* PR comments

* factor out timelocknamespaces

* Dollar sign did not help

* Revert bogus change

* test

* Fixes