Use structured locks for observability
PaxosStateLogImpl and PaxosLeaderElectionService use structured
locks (the synchronized keyword) for better observability.
This makes it easier to discover which thread is holding locks
while others wait.

Structured locks allow the jit to perform optimziations that
aren't possible for unstructured locks, for example lock coarsening[1]
and lock elision[2], on top of generally better performance of
`synchronized` on modern JVMs.

1. https://shipilev.net/jvm/anatomy-quarks/1-lock-coarsening-for-loops/
2. https://shipilev.net/jvm/anatomy-quarks/19-lock-elision/