add refresh failure callbacks
Adds an option to register callback functions that get called whenever a
LockRefreshingLockService stops refreshing a lock because it failed to
refresh.

This helps with situations where a client intentionally will be holding
locks for a long time, and they want to be nofied immediately as soon as
refreshing the lock fails. Without this callback, clients would need to
implement their own background thread to check if the locks were still
valid.