Merge some per-txn-manager thread pools into shared thread pools
As of now, the thread pools I merge are responsible for around 480
threads on production stacks I've looked at. Each of them is doing some
menial task that probably doesn't need to be done in parallel, but needs
scheduling.

Shared scheduled thread pools are hard to get right, so instead we just
share between tasks with the same type. We therefore take 8 thread pools
(480 threads) and replace with 8 threads.

Given the durations of tasks, this shouldn't affect top end performance
a whole lot.