remove unnecessary oracle config restrictions (#4988)
* remove unnecessary oracle config restrictions

Forbidding configuring the keystore/truststore when the oracle protocol
is not TCPS makes it a lot harder to automate configs. It is much easier
to allow our config automation to always fill in these values even when
they won't be used.

This does not remove the validation that checks they are present when
they really are required.

* Add generated changelog entries