only log at debug for unused keystore/truststore (#4994)
* only log at debug for unused keystore/truststore

Changes logging to only log at debug instead of warn, since it is normal
for our internal deployment infrastructure to always fill in values.

Using warn here would is generally not useful since if the config is
correct we do not want to confuse people. Either the config will be
correct and connections to the DB will work, or they won't. If someone
made a "mistake" trying to set up TCPS, it would be obvious since
connections would not work at all.

* Add generated changelog entries