Prototype of getSortedColumns
Internally we do a fair amount of merge sorting between rows, and this
is quite inefficient within Atlas. How it goes wrong is related to the
batching - basically each batch is lock checked, timestamp checked
independently, and so you get the nice waterfall in the gantt chart.

Here we add a specific method that does the sorting and merging, giving
the user an iterator which broadly does the right thing.

Skip all the rebatching stuff because sweep works nicely now.

Obviously I have no tests, and there are quite a few that you'd need to
write - especially since I've caught three more bugs in the original
getColumnsRange implementation (unsafe usages of transformEntries).

The other question I have is: most use cases (although not all) for this
actually read a fixed number of things. We could probably simplify the
contract a bit (although the implementation is already very clean) by
returning a Map of the first n or whatever.

Lastly, there is some downside of this. The present approach means that
if you e.g. read _all_ the columns of 1000 rows using distinct getColumnRange
calls, they'll be batched together at conflict checking. This approach
cannot extend in this way, so if you make 1000 calls, you'll do 1000
waterfalls here. That said, based on use cases this is probably fine -
the fanouts we see _should_ be on gets.

It does mean that if one row is requested, we simply delegate to the
present method of read only conflict checking since this will have
more opportunities to be optimized in conflict checking.

I think long term there's some fairly strong argument here that we're
suffering due to our KVS interface. IRL, the mgms implementation if
exposed on KVS would make the conflict checking here be awesome.