add limited cache warming option to ssm
Adds a CacheWarming option that can control how the SweepStrategyManager
cache is warmed on initialization. The possible options are FULL,
LIMITED, and NONE. FULL remains the default (so default behavior is not
changing). LIMITED causes only a max of 50 tables to be used for warming
the cache. NONE skips warming the cache.