add cache warming option to ssm (#5105)
* add limited cache warming option to ssm

Adds a CacheWarming option that can control how the SweepStrategyManager
cache is warmed on initialization. The possible options are FULL,
LIMITED, and NONE. FULL remains the default (so default behavior is not
changing). LIMITED causes only a max of 50 tables to be used for warming
the cache. NONE skips warming the cache.

* Add generated changelog entries

* fix limited metadata tests for cassandra

The Cassandra integration tests leave table metadata in an inconsistent
state which was causing the new test to fail. Adds a workaround where
all tables are dropped at the start of this test.

* remove getLimited versions of get tables/metadata

To avoid complexity, reducing change to just support FULL and NONE as
possible options for cache warming. If desired we could implement
limited versions later with streams etc, but NONE should fix the
internal issue.