remove moreutils dependency in build
This was just adding timestamps to the build output, but the way it was
installing moreutils is extremely unreliable (the url to download
changes frequently). Additionally if you tried using apt-get to install
the package, it would get a version that did not support the options
being used (so that just doesn't work). Any value we were getting from
this likely was completely negated by the fact that builds would break
so often, and this is one of the blockers that makes backporting bug
fixes more difficult.