Avoid creating thousands of get-ranges threads
Our metrics show services with thousands of threads in the
`serializabletransactionmanager-get-ranges` pool, however the
executor was not instrumented with Tritium, so it's not clear
how saturated it is, or if it is used at all.

Using PTExecutors factories we get tracing and execution metrics
for free, as well as resource utilization improvements to share
a slice of an underlying cached executor so threads are only
used as needed. The provided `numThreads` is still an upper limit
for the ExecutorService instance, however idle threads can be
used elsewhere.

The existing queue size warning logic is preserved using a counter
rather than instrumenting the queue itself in much the same way
tritium estimates ExecutorService queue size.