update ojdbc driver to 21.1.0.0 (#5228)
* Add generated changelog entries

* update ojdbc driver to 21.1.0.0

Also fixes an incorrect use of wildcards in versions.props for this dep.
The different deps in this group do not all have the same set of
versions available, so this should not use wildcards. It would also
probably cause problems if we had multiple dependencies coming from this
same group (e.g. you should have only one of ojdbc8 or ojdbc11; not
both).

* Add generated changelog entries

* Add generated changelog entries