allow hikari password changes at runtime
This adds a method to the hikari connection manager so that the password
can be changed at runtime. This is a requirement to support being able
to rotate database passwords without needing to restart services.

Also fixes a bug that made HikariCpConnectionManagerTest not actually
run (because it was not in a suite). The test itself was actually broken
because it did not wait for the DB to be ready (the suite setup
correctly makes sure that it can connect to the DB before running
tests).