add dbkvs runtime password config
This adds support for live reloading DB passwords for products that use
the standard install and runtime configs for Atlas. Note that previously
there was no dbkvs runtime config, so that has been added. When the
dbkvs runtime config is absent, behavior is identical to before (the
password from the install config is used).