Make CachedTransformingSupplier exception safe
If `outputFunction.apply` throws an exception, the supplier will be
stuck in a broken state until the input changes again. Swapping the
order of assignment will only update the input if the output computation
succeeds, forcing the supplier to recompute the output every call until
it does not error.

This makes it so that errors in `outputFunction` show up in error
stacktraces, instead of getting a mysterious `NullPointerException`
after the first failure when calling `CachedTransformingSupplier#get`.