[Runtime config] Part 2c: Wiring Cassandra Client Factory (#6010)
`CassandraClientFactory` now uses a narrowed `CassandraKeyValueServiceConfig` called `CassandraClientConfig`. 

As a result of `keyspace` now being required for construction, both `CassandraService` and `CassandraClientFactory` now throw at construction if `keyspace` is not present. 

Previously, `CassandraService` would throw if `keyspace` is missing but `servers` is not empty, and `CassandraClientFactory` would throw on each call of `create` if `keyspace` is missing.