Upgrade JDBI to 3.30.0
Antlr 4 recently introduced a breaking change in 4.10, meaning
libraries that generated parsers with a version before that change
can't be used with libraries that generated parsers after the
change.

JDBI was one of these libraries with a pre-4.10 version,
so in 3.28.0 they switched to using a shaded version included in
their jar. Upgrade to the current version to avoid blocking upgrades
of other Antlr-using libraries.