Oracle truncate should release space
Without DROP STORAGE, space consumed by the table remains allocated (but free) to the table and isn't released back to the tablespace.

This is problematic for applications that balloon a table (eg for initial data ingest) but all subsequent use it much smaller scale - we end up with a tonne of wasted space.

The only downside is that "DROP STORAGE" will mean if the truncated table actually needs to grow large again, there is a little more work to be done by Oracle to allocate new extents. Relatively speaking, this is minor.