Choose optimal IOT compression based on table params (#6498)
Determine the optimal Index Organized Table compression setting.

When enabling compression on an IOT, we can instruct Oracle to compress at most one fewer columns than the number of columns that make up the PK for the IOT (i.e. the IOT we create have a PK of {row, col, ts} so we can compress 0, 1 or 2 columns - nothing, row, or row & col. It wouldn't make sense to compress all three columns, because necessarily there can be multiple rows that share all three values.

When compression is enabled in an IOT, the block header adds a table the captures the values of the compressed columns. The actual rows in the blocks then refer back to entries in the header table to capture their values for the compressed columns. This extra header table and the slot references for data rows consume space in the block, so misuse of compression results in extra storage being consumed.

Compression only makes sense when we expect that the { row } or { row, col } values will be reused:

 - If the table's sweep strategy is CONSERVATIVE, then every { row, col } will have at least two rows, the current value (which may have null val if the value is deleted) and a sentinel row with ts = -1. Clearly compression of both { row, col } is valuable.

 - If the sweep strategy is not CONSERVATIVE,  there will be reuse of the { row } if there is more than one named column, or the table uses dynamic columns, which would be generally be expected to have multiple row values. In this case, compression 1 is appropriate so { row } only needs to be store once.

 - Finally, if the sweep strategy is not CONSERVATIVE, there are no sentinels, and if there is only one named column then there will be no rows that share { row } so compression should be disabled.