Log UserAgent for new timelock namespaces
Usually UserAgent appears in the request logs for a service,
however TimeLock receives so many requests that it often has
request logging disabled.

It can sometimes be hard to match up a TimeLock namespace with
the service its requests come from. Thus, this enables logging
the UserAgent for any requests that are for new (at least new
in memory) TimeLock client namespaces.

It's best effort so it handles Undertow interfaces, which is what
should be used in production everywhere anyways.