[Antithesis] Run workloads in a random order + allow running of just one workload (#6919)
The workload server can be configured to have an `executionConfig` with a `runMode` of either `ALL` or `ONE`. The `ALL` mode runs all workflows in a random order (that pokes out to the system random number generation); the `ONE` mode runs one workflow (that also is selected by random number generation).