Use Async cache to avoid cache contention
Several stack trace from our service shows blocked threads on this cache when trying to getRows. Generally these are during service restart as the cache is cold.

The async loading cache will have internally store a CompletableFuture in the ConcurrentHashMap to avoid locking inside the ConcurrentHashMap and blocking concurrent accesses to nodes in the same bucket.

Initial capacity impacts the number of buckets in the internal ConcurrentHashMap used by the cache. We want to increase the minimum to reduce contention under heavy concurrent access. This will avoid situations where threads that ask for cache values are blocked due to resizing of the underlying hashMap