remove pull-all-images.sh
If you watch the circle build, this script just looped forever failing
to pull the 1science/nginx image. The actual nginx image used in tests
is a completely different image. This script doesn't appear to actually
be useful, and just adds thousands of lines of output (repeated errors)
to every circle build.